app/backend/requirements.txt
fastapi>=0.114.0
uvicorn>=0.30.6
pytest>=8.3.2
sqlalchemy>=2.0.35
pymysql>=1.1.1
cryptography>=43.0.1
python-dotenv>=1.0.1
faker>=29.0.0
httpx>=0.27.2
python-multipart>=0.0.9
